+++ docs/spec/src/index.ts
@@ -1,9 +1,18 @@
 import { IntegrationSpecConfig } from '@jupiterone/integration-sdk-core';
 
 import { IntegrationConfig } from '../../../src/config';
-import { accessSpec } from './access';
-import { accountSpec } from './account';
+import { organizationSpec } from './organization';
+import { databaseSpec } from './database';
+import { ipAddressSpec } from './ip-address';
+import { userSpec } from './user';
+import { accessRoleSpec } from './access-role';
 
 export const invocationConfig: IntegrationSpecConfig<IntegrationConfig> = {
-  integrationSteps: [...accountSpec, ...accessSpec],
+  integrationSteps: [
+    ...organizationSpec,
+    ...databaseSpec,
+    ...ipAddressSpec,
+    ...userSpec,
+    ...accessRoleSpec,
+  ],
 };