+++ src/steps/organization/converter.ts
@@ -1,0 +1,26 @@
+import {
+  createIntegrationEntity,
+  Entity,
+} from '@jupiterone/integration-sdk-core';
+
+import { Entities } from '../constants';
+
+export function getOrganizationKey(name: string): string {
+  return `hexnode_organization:${name.replace(' ', '_')}`;
+}
+
+export function createOrganizationEntity(organization: {
+  name: string;
+}): Entity {
+  return createIntegrationEntity({
+    entityData: {
+      source: organization,
+      assign: {
+        _key: getOrganizationKey(organization.name),
+        _type: Entities.ORGANIZATION._type,
+        _class: Entities.ORGANIZATION._class,
+        name: organization.name,
+      },
+    },
+  });
+}